refactor(epic-1973): extract zeph-config crate from zeph-core god crate
Phase 1a: Create new zeph-config (Layer 1) crate with all pure-data config types.

Changes:
- Create crates/zeph-config/ with 16 modules (config, orchestration, subagent, sanitizer, features, etc.)
- Move pure-data types: SecurityConfig, RateLimitConfig, TaskEvent, TaskId, TaskOutcome, MemoryScope, PermissionMode, etc.
- zeph-core re-exports all types at original paths (backward compatible)
- Delete 9 orphaned type definition files from zeph-core/src/config/types/
- Widen escape_delimiter_tags, apply_spotlight to pub (required by agent/context/assembly.rs)
- Add SubAgentHandle::for_test() helper for agent tests
- Update root Cargo.toml to include zeph-config in workspace

Test results: 6,160 tests pass, 0 failures, feature gates preserved.

Closes #1973 Phase 1a
