+++ Cargo.toml
@@ -116,6 +116,7 @@ zeph-a2a = { path = "crates/zeph-a2a", version = "0.15.3" }
 zeph-acp = { path = "crates/zeph-acp", version = "0.15.3" }
 zeph-channels = { path = "crates/zeph-channels", version = "0.15.3" }
 zeph-common = { path = "crates/zeph-common", version = "0.15.3" }
+zeph-config = { path = "crates/zeph-config", version = "0.15.3" }
 zeph-core = { path = "crates/zeph-core", version = "0.15.3" }
 zeph-gateway = { path = "crates/zeph-gateway", version = "0.15.3" }
 zeph-index = { path = "crates/zeph-index", version = "0.15.3" }
